requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
pytest>=7.4.0
matplotlib>=3.7.0
seaborn>=0.12.0
